# docs/api/openapi.generated.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/households": {
      "post": {
        "operationId": "createHousehold",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "8-200 character opaque key reused only for retries of the same logical command",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Household created with the caller as ACTIVE OWNER",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "Create a household as its owner",
        "tags": [
          "Households"
        ]
      },
      "get": {
        "operationId": "listHouseholds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active household memberships",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseholdListItemDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "List active household memberships",
        "tags": [
          "Households"
        ]
      }
    },
    "/v1/households/{householdId}": {
      "get": {
        "operationId": "getHousehold",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Household visible to the active member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "Read an active household",
        "tags": [
          "Households"
        ]
      },
      "patch": {
        "operationId": "updateHousehold",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated household settings and incremented version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "Update owner-managed settings",
        "tags": [
          "Households"
        ]
      }
    },
    "/v1/households/{householdId}/members": {
      "get": {
        "operationId": "listMembers",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active household members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "List active household members",
        "tags": [
          "Households"
        ]
      }
    },
    "/v1/households/{householdId}/invitations": {
      "post": {
        "operationId": "createInvitation",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "8-200 character opaque key reused only for retries of the same logical command",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          },
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pending invitation with its one-time raw token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "Create a single-use invitation",
        "tags": [
          "Households"
        ]
      }
    },
    "/v1/households/{householdId}/ownership-transfer": {
      "post": {
        "operationId": "transferOwnership",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "8-200 character opaque key reused only for retries of the same logical command",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          },
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atomic old-owner/new-owner membership projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipTransferResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "Transfer ownership atomically",
        "tags": [
          "Households"
        ]
      }
    },
    "/v1/households/{householdId}/members/{memberId}": {
      "delete": {
        "operationId": "removeMember",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "8-200 character opaque key reused only for retries of the same logical command",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical membership transitioned to REMOVED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "Remove a member without deleting history",
        "tags": [
          "Households"
        ]
      }
    },
...
  },
  "info": {
    "title": "TurnTable API",
    "description": "TurnTable API-first modular monolith implementation contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "oidcBearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "OIDC access token"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "display_name": {
            "type": "string",
            "example": "Alex Rivera"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          }
        },
        "required": [
          "id",
          "display_name"
        ]
      },
...
      "ProblemErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://api.turntable.example/problems/validation-error"
          },
          "title": {
            "type": "string",
            "example": "Request validation failed"
          },
          "status": {
            "type": "number",
            "example": 422
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "detail": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemErrorDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "trace_id"
        ]
      },
...
      "CreateHouseholdRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "example": "Household Alpha"
          },
          "timezone": {
            "type": "string",
            "example": "America/New_York"
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "example": "USD"
          }
        },
        "required": [
          "name",
          "timezone",
          "currency"
        ]
      },
      "HouseholdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "example": "Household Alpha"
          },
          "timezone": {
            "type": "string",
            "example": "America/New_York"
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "example": "USD"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "id",
          "name",
          "timezone",
          "currency",
          "version"
        ]
      },
      "MembershipAccessDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE"
            ]
          }
        },
        "required": [
          "id",
          "role",
          "status"
        ]
      },
      "HouseholdListItemDto": {
        "type": "object",
        "properties": {
          "household": {
            "$ref": "#/components/schemas/HouseholdDto"
          },
          "membership": {
            "$ref": "#/components/schemas/MembershipAccessDto"
          }
        },
        "required": [
          "household",
          "membership"
        ]
      },
      "UpdateHouseholdRequestDto": {
        "type": "object",
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "timezone": {
            "type": "string",
            "example": "America/Chicago"
          }
        },
        "required": [
          "expected_version"
        ]
      },
      "MemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "household_id": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "LEFT",
              "REMOVED"
            ]
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "left_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "id",
          "household_id",
          "user",
          "role",
          "status",
          "joined_at",
          "left_at",
          "version"
        ]
      },
      "CreateInvitationRequestDto": {
        "type": "object",
        "properties": {
          "invited_email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "expires_in_hours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 168,
            "default": 24
          }
        }
      },
      "InvitationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "household_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING"
            ]
          },
          "invited_email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "token": {
            "type": "string",
            "description": "Single-use opaque token returned only by invitation creation",
            "minLength": 32
          }
        },
        "required": [
          "id",
          "household_id",
          "status",
          "invited_email",
          "expires_at",
          "token"
        ]
      },
      "TransferOwnershipRequestDto": {
        "type": "object",
        "properties": {
          "target_member_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "target_member_id"
        ]
      },
      "OwnershipTransferResponseDto": {
        "type": "object",
        "properties": {
          "household_id": {
            "type": "string",
            "format": "uuid"
          },
          "previous_owner": {
            "$ref": "#/components/schemas/MemberDto"
          },
          "new_owner": {
            "$ref": "#/components/schemas/MemberDto"
          }
        },
        "required": [
          "household_id",
          "previous_owner",
          "new_owner"
        ]
      }
    }
  }
}